third_party/blink/web_tests/external/wpt/css/css-values/calc-size/calc-size-width-expected.txt

This is a testharness.js-based test.
[FAIL] resolved width for width: calc-size(stretch, size / 2 - 10%)
  assert_equals: expected "200px" but got "500px"
Harness: the test ran to completion.
